First project
Create a List of strings and add numbers to that list while using a for loop. Use a string type in order to type "done" after adding the numbers. After "done" is typed use an if statement and sort function to sort the string of numbers in the list from smallest to largest. Make sure to remove the "done" string in the list so it doesn't appear after using a foreach loop to print the sorted numbers after that.