Using and practicing on "Lists" - finding the average of all elements
Initiating with an integer of sum to store the values. Opening a while statement to run. Inputting a string and adding it to an if statement. Using a parse method to check if the string can be parsed into an int. If yes, to add that value to the list as an int. After it is added sum takes the value of every number typed and prints the sum of the numbers. When "done" is typed the last result if the total sum of all the numbers divided by their count.